fix docker setup and upgrade buildbot to 2.5
The previous docker setup was slightly incomplete, this should simplify
the basic setup as it only requires to run `docker-compose up`,
automatically building all required containers.

Nginx with letsencrypt is used for the *proof-of-concept* deployment.
This also allows to encrpyt the traffic between remote worker using
Nginx streams with ssl.

The new Buildbot 2.5 has a nicer web interface and likely many other
features to explore and optimize, this is however only a first committo
move to version 2.5.

Signed-off-by: Paul Spooren <mail@aparcar.org>