phase1: cleanup
ConfigParser .get() can use `fallback` to reduce the number of temporary
defined variables, make use of that.

Also automatically set the working directory to unshared if there are
concurrent builds.

Remove generic comments to be replaced with OpenWrt specific
information.

Replace trivial code bloat like "if foo is not None" with "if foo"

Signed-off-by: Paul Spooren <mail@aparcar.org>