ability to add external targets
This patch allows to add external targets via the `config.ini`.

To to so a new section `external_targets` in the following format can be
used:

    [external_targets]
    special_mips = git-src,https://git.example.org/special_mips.git

The master process automatically updates and install the specified feeds
and add them as builders.

Additionally the option `external_targets_only` in section `phase1` is
added to *only* build the targets added via a feed.

Signed-off-by: Paul Spooren <mail@aparcar.org>