buildbot: ability to build single profiles
This patch adds the ability to add profiles to the `active_profiles`
sections to build them based on a `gen_config.sh` configuration.

As the worker uses a shared buildroot the external feeds are globally
installed, however only activated based on the generated `.config`.

If enabled this patch changes the builder name from `target/subtarget`
to `target/subtarget/profile`.

The example `config.ini` contains the setup (without passwords) for the
`rax40` profile.

Signed-off-by: Paul Spooren <mail@aparcar.org>