phase1: add postgresql support
The buildbot has over 50 builders and storing all results in a single
sqlite database slows down the web interface and even builds.

This patch adds a PostgreSQL container to the docker-compose file and
automatically connects to it.

During testing this greatly improved the webinterface speeds and even
speeded up the build of targets by 10-15 minutes.

Signed-off-by: Paul Spooren <mail@aparcar.org>